Remove PageRequest.java

+++ src/main/java/com/bryan/system/mapper/UserMapper.java
@@ -23,8 +23,8 @@ public interface UserMapper {
 
     SysUser selectByUsername(String username);
 
-    List<SysUser> selectPage(@Param("offset") long offset,
-                             @Param("pageSize") long pageSize,
+    List<SysUser> selectPage(@Param("offset") int offset,
+                             @Param("pageSize") int pageSize,
                              @Param("req") UserSearchRequest search,
                              @Param("export") UserExportRequest export);
